fix(compiler): simplify generated Int32 conversions
Generated Go wrapped values that were already native int32 expressions in another int32 conversion. This change widens those proven expressions directly to Go int, which removes redundant conversion lint findings without changing Haxe Int32 behavior.\n\nThe shared file runtime now documents why its ordinary user-visible file modes intentionally honor the process umask. A focused contract protects both lint outcomes, and committed snapshots and examples are refreshed from the compiler output.\n\nVerified with the changed and full test suites, 332 generated snapshots, semantic comparison, strict standard-library sweep, all examples, focused Go lint checks, both performance review harnesses, and the repository pre-commit security checks.

+++ test/snapshot/core/array_constructor_new/intended/main.go
@@ -20,7 +20,7 @@ func main() {
 	for _g < nums.Len() {
 		n := hxrt.IntFromNullableAny(nums.Get(_g))
 		_g = int(int32((_g + 1)))
-		sum = int(int32((hxrt.Int32Wrap(sum) + hxrt.Int32Wrap(n))))
+		sum = int((hxrt.Int32Wrap(sum) + hxrt.Int32Wrap(n)))
 	}
 	hxrt.Println(any(sum))
 }
