+++ test/snapshot/core/closures_capture/intended/main.go
@@ -5,7 +5,7 @@ import "snapshot/hxrt"
 func main() {
 	factor := 3
 	mul := func(v int) int {
-		return int(int32((hxrt.Int32Wrap(v) * hxrt.Int32Wrap(factor))))
+		return int((hxrt.Int32Wrap(v) * hxrt.Int32Wrap(factor)))
 	}
 	factor = 4
 	var v any = any(mul(2))
