+++ test/snapshot/core/loop_break_continue/intended/main.go
@@ -7,13 +7,13 @@ func main() {
 	sum := 0
 	for i < 10 {
 		i = int(int32((i + 1)))
-		if int(int32((hxrt.Int32Wrap(i) % hxrt.Int32Wrap(2)))) == 0 {
+		if int((hxrt.Int32Wrap(i) % hxrt.Int32Wrap(2))) == 0 {
 			continue
 		}
 		if i > 7 {
 			break
 		}
-		sum = int(int32((hxrt.Int32Wrap(sum) + hxrt.Int32Wrap(i))))
+		sum = int((hxrt.Int32Wrap(sum) + hxrt.Int32Wrap(i)))
 	}
 	var v any = any(hxrt.StdString(sum))
 	hxrt.Println(v)
