+++ test/snapshot/core/optimizer_plan_string_fastpath_enabled/intended/main.go
@@ -5,6 +5,6 @@ import "snapshot/hxrt"
 func main() {
 	s := hxrt.StringFromLiteral("héllo")
 	out := hxrt.StringConcatStringPtr(hxrt.StringConcatStringPtr(hxrt.StringConcatStringPtr(hxrt.StringCharAtStringPtr(s, 1), hxrt.StdString(hxrt.StringCharCodeAtAnyStringPtr(s, 1))), hxrt.StringSubstringStringPtr(s, 0, 3)), hxrt.StringSubstrStringPtr(s, -2, 0, false))
-	var v any = any(int(int32((hxrt.Int32Wrap(hxrt.StringLengthStringPtr(s)) + hxrt.Int32Wrap(hxrt.StringLengthStringPtr(out))))))
+	var v any = any(int((hxrt.Int32Wrap(hxrt.StringLengthStringPtr(s)) + hxrt.Int32Wrap(hxrt.StringLengthStringPtr(out)))))
 	hxrt.Println(v)
 }
